Removes unused imports and put cardlist items inside one file
