--- app/src/main/java/com/bootcamp/concrete/magicdeck/data/domain/CardListHeader.kt
@@ -1,5 +1,0 @@
-package com.bootcamp.concrete.magicdeck.data.domain
-
-
-class CardListHeader(var text: String) : CardListItem
-
