+++ app/src/main/java/com/bootcamp/concrete/magicdeck/data/domain/CardListItem.kt
@@ -1,3 +1,17 @@
 package com.bootcamp.concrete.magicdeck.data.domain
 
-interface CardListItem
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+sealed class CardListItem
+
+@Parcelize
+class Card (
+    val name: String,
+    val types: List<String>,
+    val set: String,
+    val id: String,
+    val imageUrl: String?
+): Parcelable, CardListItem()
+
+class CardListHeader(var text: String) : CardListItem()
